feat: Controller/ 특정 통증 부위 다음 질문 조회
