+++ src/main/java/com/healthier/headachelogic/service/QuestionService.java
@@ -4,6 +4,7 @@ import com.healthier.headachelogic.domain.Answer;
 import com.healthier.headachelogic.domain.Question;
 import com.healthier.headachelogic.dto.QuestionDto;
 import com.healthier.headachelogic.dto.ResultDto;
+import com.healthier.headachelogic.dto.painArea.HeadachePainAreaNextResponse;
 import com.healthier.headachelogic.repository.QuestionRepository;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
@@ -46,28 +47,6 @@ public class QuestionService {
         //다음 질문이 존재하지 않을 때
         else {
             return new HeadachePainAreaNextResponse(answer.getResultId(), answer.getResult());
-        }
-    }
-
-    @Data
-    public class HeadachePainAreaNextResponse {
-        private int type;
-        private List<QuestionDto> questions = new ArrayList<>();
-        ResultDto resultDto;
-
-        //생성자
-        protected HeadachePainAreaNextResponse() {}
-
-        //1) 생성자 : 다음 질문 반환
-        public HeadachePainAreaNextResponse(Question question) {
-            type = 1;
-            questions.add(new QuestionDto(question));
-        }
-
-        //2) 생성자 : 진단 결과 안내
-        public HeadachePainAreaNextResponse(int resultId, String result) {
-            type = 2;
-            resultDto = new ResultDto(resultId, result);
         }
     }
 }
